[BE] REFACTOR: 블랙홀매니저 최종 - 자동갱신형태로
