+++ backend/src/test/java/org/ftclub/cabinet/utils/blackhole/manager/BlackholeManagerUnitTest.java
@@ -14,18 +14,12 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import java.time.LocalDateTime;
-import java.util.List;
 import org.ftclub.cabinet.auth.service.FtApiManager;
 import org.ftclub.cabinet.config.FtApiProperties;
 import org.ftclub.cabinet.dto.UserBlackholeInfoDto;
-import org.ftclub.cabinet.exception.ExceptionStatus;
-import org.ftclub.cabinet.exception.ServiceException;
 import org.ftclub.cabinet.lent.service.LentService;
 import org.ftclub.cabinet.user.service.UserService;
-import org.ftclub.cabinet.utils.ExceptionUtil;
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -40,8 +34,6 @@ public class BlackholeManagerUnitTest {
 
 	@InjectMocks
 	private BlackholeManager blackholeManager;
-	@Mock
-	private BlackholeRefresher blackholeRefresher;
 	@Mock
 	private FtApiManager ftApiManager = mock(FtApiManager.class);
 	@Mock
@@ -88,7 +80,7 @@ public class BlackholeManagerUnitTest {
 				LocalDateTime.now().minusDays(1));
 		JsonNode JsonafterBlackholedAt = objectMapper.readTree(StringAfterBlackholedAt);
 
-		given(blackholeRefresher.getBlackholeInfo(userBlackholeInfoDto.getName())).willReturn(jsonUserInfo);
+		given(ftApiManager.getFtUsersInfoByName(name)).willReturn(jsonUserInfo);
 		given(jsonUserInfo.get("cursus_users")).willReturn(mockCursusUsers);
 		lenient().when(mockCursusUsers.get(0)).thenReturn(mockFieldZero);
 		lenient().when(mockCursusUsers.get(1)).thenReturn(mockFieldOne);
@@ -120,10 +112,9 @@ public class BlackholeManagerUnitTest {
 				beforeBlackholedAt
 		);
 
-		given(blackholeRefresher.getBlackholeInfo(userBlackholeInfoDto.getName())).willThrow(
-				new HttpClientErrorException(
-						HttpStatus.NOT_FOUND
-				));
+		given(ftApiManager.getFtUsersInfoByName(name)).willThrow(new HttpClientErrorException(
+				HttpStatus.NOT_FOUND
+		));
 
 		blackholeManager.handleBlackhole(userBlackholeInfoDto);
 
@@ -153,8 +144,7 @@ public class BlackholeManagerUnitTest {
 		JsonNode mockFieldZero = mock(JsonNode.class);
 		JsonNode mockFieldOne = mock(JsonNode.class);
 
-		given(blackholeRefresher.getBlackholeInfo(userBlackholeInfoDto.getName())).willReturn(
-				jsonUserInfo);
+		given(ftApiManager.getFtUsersInfoByName(name)).willReturn(jsonUserInfo);
 		given(jsonUserInfo.get("cursus_users")).willReturn(mockCursusUsers);
 		lenient().when(mockCursusUsers.get(0)).thenReturn(mockFieldZero);
 		given(mockCursusUsers.size()).willReturn(1);
@@ -195,8 +185,7 @@ public class BlackholeManagerUnitTest {
 				LocalDateTime.now().plusDays(42));
 		JsonNode JsonafterBlackholedAt = objectMapper.readTree(StringAfterBlackholedAt);
 
-		given(blackholeRefresher.getBlackholeInfo(userBlackholeInfoDto.getName())).willReturn(
-				jsonUserInfo);
+		given(ftApiManager.getFtUsersInfoByName(name)).willReturn(jsonUserInfo);
 		given(jsonUserInfo.get("cursus_users")).willReturn(mockCursusUsers);
 		lenient().when(mockCursusUsers.get(0)).thenReturn(mockFieldZero);
 		lenient().when(mockCursusUsers.get(1)).thenReturn(mockFieldOne);
@@ -236,8 +225,7 @@ public class BlackholeManagerUnitTest {
 //		새 블랙홀 날짜 = null
 		JsonNode JsonafterBlackholedAt = null;
 
-		given(blackholeRefresher.getBlackholeInfo(userBlackholeInfoDto.getName())).willReturn(
-				jsonUserInfo);
+		given(ftApiManager.getFtUsersInfoByName(name)).willReturn(jsonUserInfo);
 		given(jsonUserInfo.get("cursus_users")).willReturn(mockCursusUsers);
 		lenient().when(mockCursusUsers.get(0)).thenReturn(mockFieldZero);
 		lenient().when(mockCursusUsers.get(1)).thenReturn(mockFieldOne);
@@ -276,8 +264,7 @@ public class BlackholeManagerUnitTest {
 //		새 블랙홀 날짜 = null
 		JsonNode JsonafterBlackholedAt = null;
 
-		given(blackholeRefresher.getBlackholeInfo(userBlackholeInfoDto.getName())).willReturn(
-				jsonUserInfo);
+		given(ftApiManager.getFtUsersInfoByName(name)).willReturn(jsonUserInfo);
 		given(jsonUserInfo.get("cursus_users")).willReturn(mockCursusUsers);
 		lenient().when(mockCursusUsers.get(0)).thenReturn(mockFieldZero);
 		lenient().when(mockCursusUsers.get(1)).thenReturn(mockFieldOne);
@@ -316,8 +303,7 @@ public class BlackholeManagerUnitTest {
 //		새 블랙홀 날짜 = null
 		JsonNode JsonafterBlackholedAt = null;
 
-		given(blackholeRefresher.getBlackholeInfo(userBlackholeInfoDto.getName())).willReturn(
-				jsonUserInfo);
+		given(ftApiManager.getFtUsersInfoByName(name)).willReturn(jsonUserInfo);
 		given(jsonUserInfo.get("cursus_users")).willReturn(mockCursusUsers);
 		lenient().when(mockCursusUsers.get(0)).thenReturn(mockFieldZero);
 		lenient().when(mockCursusUsers.get(1)).thenReturn(mockFieldOne);
